ota|merge

+++ ventilator/src/main/java/com/robam/ventilator/ui/activity/SaleServiceActivity.java
@@ -14,11 +14,14 @@ import com.robam.common.device.Plat;
 import com.robam.common.http.DownloadListener;
 import com.robam.common.http.RetrofitCallback;
 import com.robam.common.ui.dialog.IDialog;
+import com.robam.common.utils.DeviceUtils;
 import com.robam.common.utils.LogUtils;
+import com.robam.common.utils.ToastUtils;
 import com.robam.ventilator.BuildConfig;
 import com.robam.ventilator.R;
 import com.robam.ventilator.base.VentilatorBaseActivity;
 import com.robam.ventilator.constant.DialogConstant;
+import com.robam.ventilator.device.HomeVentilator;
 import com.robam.ventilator.factory.VentilatorDialogFactory;
 import com.robam.ventilator.http.CloudHelper;
 import com.robam.ventilator.response.AppTypeRes;
@@ -36,6 +39,9 @@ public class SaleServiceActivity extends VentilatorBaseActivity {
     private String versionUrl;
     private IDialog updateDialog, progressDialog;
     int curProgress = 0;
+    String FIRMWARE = "firmware";
+    String APK = "apk";
+    private String updateMode = FIRMWARE;
 
     @Override
     protected int getLayoutId() {
@@ -67,18 +73,27 @@ public class SaleServiceActivity extends VentilatorBaseActivity {
     }
     //检查固件
     private void checkFirmware() {
-        CloudHelper.checkAppVersion(this, "RKPAD", Plat.getPlatform().getDt(), "firmware", AppTypeRes.class, new RetrofitCallback<AppTypeRes>(){
+        CloudHelper.checkAppVersion(this, "RKPAD", Plat.getPlatform().getDt(), FIRMWARE, AppTypeRes.class, new RetrofitCallback<AppTypeRes>(){
 
             @Override
             public void onSuccess(AppTypeRes appTypeRes) {
                 if (null != appTypeRes && null != appTypeRes.ver && null != appTypeRes.ver.url) {
-                    LogUtils.e(appTypeRes.ver.url);
-                    versionUrl = appTypeRes.ver.url;
-                    String version = String.format(getString(R.string.ventilator_new_version), appTypeRes.ver.code + "");
-                    tvNewVersion.setVisibility(View.VISIBLE);
-                    tvNewVersion.setText(version);
-                } else
-                    checkApk();
+                    String curVer = Plat.getPlatform().getFirmwareVersion();
+                    try {
+                        curVer = curVer.replace(".", "");
+                        if (Integer.parseInt(curVer) != appTypeRes.ver.code) { //版本不一致
+                            LogUtils.e(appTypeRes.ver.url);
+                            updateMode = FIRMWARE; //更新固件
+                            versionUrl = appTypeRes.ver.url;
+                            String version = String.format(getString(R.string.ventilator_new_version), appTypeRes.ver.code + "");
+                            tvNewVersion.setVisibility(View.VISIBLE);
+                            tvNewVersion.setText(version);
+                            return;
+                        }
+                    } catch (Exception e) {}
+                }
+
+                checkApk();
             }
 
             @Override
@@ -89,15 +104,22 @@ public class SaleServiceActivity extends VentilatorBaseActivity {
     }
     //检查apk
     private void checkApk() {
-        CloudHelper.checkAppVersion(this, "RKPAD", Plat.getPlatform().getDt(), "apk", AppTypeRes.class, new RetrofitCallback<AppTypeRes>() {
+        CloudHelper.checkAppVersion(this, "RKPAD", Plat.getPlatform().getDt(), APK, AppTypeRes.class, new RetrofitCallback<AppTypeRes>() {
             @Override
             public void onSuccess(AppTypeRes appTypeRes) {
                 if (null != appTypeRes && null != appTypeRes.ver && null != appTypeRes.ver.url) {
-                    LogUtils.e(appTypeRes.ver.url);
-                    versionUrl = appTypeRes.ver.url;
-                    String version = String.format(getString(R.string.ventilator_new_version), appTypeRes.ver.code + "");
-                    tvNewVersion.setVisibility(View.VISIBLE);
-                    tvNewVersion.setText(version);
+                    String curVer = DeviceUtils.getVersionName(getApplicationContext());
+                    try {
+                        curVer = curVer.replace(".", "");
+                        if (Integer.parseInt(curVer) != appTypeRes.ver.code) { //版本不一致
+                            LogUtils.e(appTypeRes.ver.url);
+                            updateMode = APK; //更新应用
+                            versionUrl = appTypeRes.ver.url;
+                            String version = String.format(getString(R.string.ventilator_new_version), appTypeRes.ver.code + "");
+                            tvNewVersion.setVisibility(View.VISIBLE);
+                            tvNewVersion.setText(version);
+                        }
+                    } catch (Exception e) {}
                 }
             }
 
@@ -186,34 +208,60 @@ public class SaleServiceActivity extends VentilatorBaseActivity {
                 @Override
                 public void onClick(View v) {
                     if (v.getId() == R.id.tv_ok) {
+                        String fileName = "update.zip";
+                        if (FIRMWARE.equals(updateMode))
+                            fileName = "update.zip";
+                        else
+                            fileName = "update.apk";
                         showProressDialog();
-                        CloudHelper.downloadFile(getApplicationContext(), versionUrl, new DownloadListener() {
+                        CloudHelper.downloadFile(SaleServiceActivity.this, getApplicationContext(), versionUrl, fileName, new DownloadListener() {
                             @Override
                             public void onProgress(int progress) {
-                                LogUtils.e("thread=" + Thread.currentThread() + " process=" + progress);
+                                LogUtils.e("thread=" + Thread.currentThread().getName() + " process=" + progress);
+                                HomeVentilator.getInstance().updateOperationTime(); //防止5分钟关机
                                 curProgress = progress;
                                 runOnUiThread(updateProgress);
                             }
 
                             @Override
                             public void onFinish(String path) {
-                                closeProgressDialog();
-                                //静默安装
-                                if (installSilent(path)) {
-                                    //安装成功，重启应用
-                                    Intent intent = new Intent();
-                                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                                    intent.setClassName(getApplicationContext(), "com.robam.androidpad.MainActivity");
-                                    startActivity(intent);
+                                runOnUiThread(new Runnable() {
+                                    @Override
+                                    public void run() {
+                                        closeProgressDialog();
+                                    }
+                                });
 
-                                    android.os.Process.killProcess(android.os.Process.myPid());
-                                    System.exit(10);
+                                HomeVentilator.getInstance().updateOperationTime();
+                                if (FIRMWARE.equals(updateMode)) {
+                                    if (Plat.getPlatform().verifyPackage(path)) {
+                                        Plat.getPlatform().intallPackage(path);
+                                    }
+                                } else {
+                                    //静默安装
+                                    if (installSilent(path)) {
+                                        //安装成功，重启应用
+                                        Intent intent = new Intent();
+                                        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                        intent.setClassName(getApplicationContext(), "com.robam.androidpad.MainActivity");
+                                        startActivity(intent);
+
+                                        android.os.Process.killProcess(android.os.Process.myPid());
+                                        System.exit(10);
+                                    }
                                 }
                             }
 
                             @Override
                             public void onFail(String errorInfo) {
-                                closeProgressDialog();
+                                HomeVentilator.getInstance().updateOperationTime();
+                                runOnUiThread(new Runnable() {
+                                    @Override
+                                    public void run() {
+                                        closeProgressDialog();
+                                        ToastUtils.showShort(getApplicationContext(), "下载失败");
+                                    }
+                                });
                             }
                         });
